test(liveliness): inline shared helper, drop support module
Per review: the shared support/mod.rs added an extra file for a helper
only two tests used, and running them in the same process (rather than
separate cargo test binaries) previously caused real cross-test
contamination. Inline the ~90-line helper into each test file directly
instead, keeping the two files self-contained and process-isolated
without a third module file.
